refactor replay metadata scanner boundaries
